Assignment 2: A* algorithm Simulation
Task 2 of the course.
Informal description: implement a program that simulates the resolution of a path-finding problem using A\* algorithm. Solution should print the states and possible transitions, as well as the costs of G and H functions on each movement. Als should have a simple animation and the grid size, percentage of obstacles, and starting and target points should be parameterized.
A NetBeans project with the solution can be found on the "Astar" folder.